LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/style-src-inline-style-nonce-blocked-expected.txt


PASS Should not load inline style element with invalid nonce
FAIL Should fire a securitypolicyviolation event assert_equals: expected "style-src 'self' 'nonce-nonceynonce'" but got "style-src-elem"
Lorem ipsum
